Lecture User#1 / User#2 / Foreign Key

+++ JssProject/main/views.py
@@ -2,6 +2,8 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .forms import JssForm
 from .models import Jasoseol
 from django.http import Http404
+from django.core.exceptions import PermissionDenied
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -9,15 +11,23 @@ def index(request):
     all_jss = Jasoseol.objects.all()
     return render(request, "index.html", {'all_jss': all_jss})
 
+def my_index(request):
+    my_jss = Jasoseol.objects.filter(author=request.user)
+    return render(request, "index.html", {'all_jss': my_jss})
+
+@login_required(login_url="/login/")
 def create(request):
     if request.method == "POST":
         filled_form = JssForm(request.POST)
         if filled_form.is_valid():
-            filled_form.save()
+            temp_form = filled_form.save(commit=False)
+            temp_form.author = request.user
+            temp_form.save()
             return redirect('index')
     jss_form = JssForm()
     return render(request, "create.html", {'jss_form':jss_form})
 
+@login_required(login_url="/login/")
 def detail(request, jss_id):
     # try:
     #     my_jss = Jasoseol.objects.get(pk=jss_id)
@@ -29,8 +39,10 @@ def detail(request, jss_id):
 
 def delete(request, jss_id):
     my_jss = Jasoseol.objects.get(pk=jss_id)
-    my_jss.delete()
-    return redirect('index')
+    if request.user == my_jss.author:
+        my_jss.delete()
+        return redirect('index')
+    raise PermissionDenied
 
 def update(request, jss_id):
     my_jss = Jasoseol.objects.get(pk=jss_id)
